Release 0.22

+++ src/game/Server/WorldSocket.cpp
@@ -220,7 +220,7 @@ int WorldSocket::open(void* a)
     return SendPacket(packet);
 }
 
-int WorldSocket::close(int)
+int WorldSocket::close(u_long)
 {
     shutdown();
